+++ src/com/genir/aitweaks/core/playerassist/AIDrone.kt
@@ -4,6 +4,7 @@ import com.fs.starfarer.api.Global
 import com.fs.starfarer.api.combat.CollisionClass
 import com.fs.starfarer.api.combat.ShipAIPlugin
 import com.fs.starfarer.api.combat.ShipAPI
+import com.genir.aitweaks.core.extensions.timeMult
 
 fun makeAIDrone(ai: ShipAIPlugin): ShipAPI {
     val spec = Global.getSettings().getHullSpec("dem_drone")
@@ -19,4 +20,17 @@ fun makeAIDrone(ai: ShipAIPlugin): ShipAPI {
     aiDrone.shipAI = ai
 
     return aiDrone
+}
+
+/** AI drone time multiplier needs to be the same as the player ship time multiplier.
+ * This is required for the game engine to advance the AI the same number of times
+ * as the player ship is advanced when the player ship is in fast-time mode. */
+fun syncTimeWithPlayerShip(aiDrone: ShipAPI) {
+    val droneTime = aiDrone.mutableStats.timeMult
+    val playerShip = Global.getCombatEngine().playerShip
+    if (playerShip != null) {
+        droneTime.modifyMult("aitweaks_ai_drone_time_sync", playerShip.timeMult)
+    } else {
+        droneTime.unmodifyMult("aitweaks_ai_drone_time_sync")
+    }
 }